+++ R/class_definitions.R
@@ -6,7 +6,7 @@ Workbook <- setRefClass("Workbook",
                         
                         fields = c("sheet_names" = "character",
                                    "workbookProtection" = "ANY",
-                                   
+                               
                                    "charts" = "ANY",
                                    "isChartSheet" = "logical",